submission-worker/requirements.txt
annotated-types==0.7.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
dnspython==2.7.0
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-core==2.4.3
google-cloud-pubsub==2.29.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
importlib_metadata==8.6.1
lazy-model==0.2.0
mongoengine==0.29.1
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic_core==2.33.2
pymongo==4.13.0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
toml==0.10.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wrapt==1.17.2
zipp==3.22.0
